+++ src/components/organisms/Navigation/Navigation.tsx
@@ -2,6 +2,7 @@ import logo from '/assets/images/logo.png';
 import { Icon, NavButton, Persona } from '@/components/atoms';
 import NavMenu from '@/components/molecules/NavMenu/NavMenu';
 import { logoutAPI } from '@/services/auth';
+import { NavItem } from './NavItem';
 
 interface NavigationProps {
     generation: string;
@@ -10,52 +11,6 @@ interface NavigationProps {
 }
 
 const Navigation = ({ generation, name }: NavigationProps) => {
-    const navItem = [
-        {
-            name: '메인페이지',
-            path: '/',
-            icon: 'calendar',
-        },
-        {
-            name: '프로젝트',
-            icon: 'document',
-            children: [
-                {
-                    name: '관통 프로젝트',
-                    path: '/track',
-                },
-                {
-                    name: '파이널 프로젝트',
-                    path: '/final',
-                },
-            ],
-        },
-        {
-            name: '학습',
-            icon: 'pen',
-            children: [
-                {
-                    name: '알고리즘',
-                    path: '/algorithm',
-                },
-                {
-                    name: '웹',
-                    path: '/web',
-                },
-            ],
-        },
-        {
-            name: '음악',
-            path: '/music',
-            icon: 'music',
-        },
-        {
-            name: '학생관리',
-            path: '/admin',
-            icon: 'edit-user',
-        },
-    ];
-
     const handleLogout = async () => {
         try {
             await logoutAPI();
@@ -73,8 +28,9 @@ const Navigation = ({ generation, name }: NavigationProps) => {
             <Persona generation={generation} name={name} />
             <nav className="h-full">
                 <ul className="flex h-full flex-col">
-                    <NavMenu menus={navItem} />
+                    <NavMenu menus={NavItem} />
                     <NavButton
+                        key="마이페이지"
                         id="마이페이지"
                         title="마이페이지"
                         icon={<Icon id="user" />}
@@ -82,6 +38,7 @@ const Navigation = ({ generation, name }: NavigationProps) => {
                         className="mt-auto"
                     />
                     <NavButton
+                        key="로그아웃"
                         id="로그아웃"
                         title="로그아웃"
                         icon={<Icon id="circle-close-red" />}
